Updated to use new html table builder in Caffeine core. Also added comments.

+++ controllers/admin_post.php
@@ -2,18 +2,15 @@
 
 class Blog_Admin_PostController extends Controller {
 
-
+    /**
+     * Displays a table of blog posts.
+     *
+     * Route: admin/blog/posts/manage
+     */
     public static function manage()
     {
-        $rows = array();
-        $header = array(
-            array(
-                'Title', 
-                'attributes' => array(
-                    'colspan' => 2
-                )
-            )
-        );
+        $table = Html::table();
+        $table->addHeader()->addCol('Title', array('colspan' => 2));
         
         $posts = Blog::post()->orderBy('created_at', 'DESC')->all();
 
@@ -21,65 +18,54 @@ class Blog_Admin_PostController extends Controller {
         {   
             foreach($posts as $p)
             {
-                $rows[] = array(
-                    Html::a()->get($p->title, 'admin/blog/posts/edit/' . $p->id),
-                    array(
-                        Html::a()->get('Delete', 'admin/blog/posts/delete/' . $p->id),
-                        'attributes' => array(
-                            'class' => 'right'
-                        )
-                    )
+                $row = $table->addRow();
+                $row->addCol(Html::a()->get($p->title, 'admin/blog/posts/edit/' . $p->id));
+                $row->addCol(
+                    Html::a()->get('Delete', 'admin/blog/posts/delete/' . $p->id),
+                    array('class' => 'right')
                 );
             }
         }
         else
-        {
-            $rows[] = array(
-                array(
-                    '<em>No posts.</em>',
-                    'attributes' => array(
-                        'colspan' => 2
-                    )
-                )
-            );
-        }
+            $table->addRow()->addCol('<em>No posts.</em>', array('colspan' => 2));
 
         return array(
             'title' => 'Manage Posts',
-            'content' => Html::table()->build($header, $rows)
+            'content' => $table->render()
         );
     }
 
-
+    /**
+     * Displays a form for creating a new blog post.
+     *
+     * Route: admin/blog/posts/create
+     */
     public static function create()
     {
-        if($_POST)
+        if(isset($_POST['create_post']) && Html::form()->validate())
         {
-            if(Html::form()->validate())
+            $postId = Blog::post()->insert(array(
+                'user_id' => User::current()->id,
+                'title' => $_POST['title'],
+                'slug' => String::slugify($_POST['title']),
+                'body' => $_POST['body']
+            ));
+
+            if($postId)
             {
-                $postId = Blog::post()->insert(array(
-                    'user_id' => User::current()->id,
-                    'title' => $_POST['title'],
-                    'slug' => String::slugify($_POST['title']),
-                    'body' => $_POST['body']
-                ));
-
-                if($postId)
+                foreach($_POST['category_id'] as $catId)
                 {
-                    foreach($_POST['category_id'] as $catId)
-                    {
-                        Db::table('categories_posts')->insert(array(
-                            'category_id' => $catId,
-                            'post_id' => $postId
-                        ));
-                    }
-
-                    Message::ok('Post created successfully.');
-                    $_POST = array(); // Clear form
+                    Db::table('categories_posts')->insert(array(
+                        'category_id' => $catId,
+                        'post_id' => $postId
+                    ));
                 }
-                else
-                    Message::error('Error creating post. Please try again.');
+
+                Message::ok('Post created successfully.');
+                $_POST = array(); // Clear form
             }
+            else
+                Message::error('Error creating post. Please try again.');
         }
 
         $formData[] = array(
@@ -91,7 +77,8 @@ class Blog_Admin_PostController extends Controller {
                 ),
                 'body' => array(
                     'title' => 'Body',
-                    'type' => 'textarea'
+                    'type' => 'textarea',
+                    'attributes' => array('class' => 'tinymce')
                 ),
                 'category_id[]' => array(
                     'title' => 'Category',
@@ -102,7 +89,7 @@ class Blog_Admin_PostController extends Controller {
                     ),
                     'validate' => array('required')
                 ),
-                'submit' => array(
+                'create_post' => array(
                     'type' => 'submit',
                     'value' => 'Create Post'
                 )
@@ -115,41 +102,45 @@ class Blog_Admin_PostController extends Controller {
         );
     }
 
-
+    /**
+     * Displays a form for editing a blog post.
+     *
+     * Route: admin/blog/posts/edit/:id
+     *
+     * @param int $id The id of the blog post to edit.
+     */
     public static function edit($id)
     {
         if(!$post = Blog::post()->find($id))
             return ERROR_NOTFOUND;
 
-        if($_POST)
+        if(isset($_POST['update_post']) && Html::form()->validate())
         {
-            if(Html::form()->validate())
+            $status = Blog::post()->where('id', '=', $id)->update(array(
+                'title' => $_POST['title'],
+                'slug' => $_POST['slug'],
+                'body' => $_POST['body']
+            ));
+
+            if($status)
             {
-                $status = Blog::post()->where('id', '=', $id)->update(array(
-                    'title' => $_POST['title'],
-                    'slug' => $_POST['slug'],
-                    'body' => $_POST['body']
-                ));
+                // Clear old categories
+                Db::table('categories_posts')->where('post_id', '=', $id)->delete();
 
-                if($status)
+                // Insert new categories
+                foreach($_POST['category_id'] as $catId)
                 {
-                    // Clear old categories
-                    Db::table('categories_posts')->where('post_id', '=', $id)->delete();
-
-                    // Insert new categories
-                    foreach($_POST['category_id'] as $catId)
-                    {
-                        Db::table('categories_posts')->insert(array(
-                            'category_id' => $catId,
-                            'post_id' => $id
-                        ));
-                    }
-
-                    Message::ok('Post updated successfully.');
+                    Db::table('categories_posts')->insert(array(
+                        'category_id' => $catId,
+                        'post_id' => $id
+                    ));
                 }
-                else
-                    Message::error('Error updating post. Please try again.');
+
+                Message::ok('Post updated successfully.');
+                $post = Blog::post()->find($id); // Get updated for form
             }
+            else
+                Message::error('Error updating post. Please try again.');
         }
 
         $formData[] = array(
@@ -169,7 +160,8 @@ class Blog_Admin_PostController extends Controller {
                 'body' => array(
                     'title' => 'Body',
                     'type' => 'textarea',
-                    'default_value' => $post->title
+                    'attributes' => array('class' => 'tinymce'),
+                    'default_value' => $post->body
                 ),
                 'category_id[]' => array(
                     'title' => 'Category',
@@ -180,7 +172,7 @@ class Blog_Admin_PostController extends Controller {
                     ),
                     'validate' => array('required')
                 ),
-                'submit' => array(
+                'update_post' => array(
                     'type' => 'submit',
                     'value' => 'Update Post'
                 )
@@ -193,7 +185,13 @@ class Blog_Admin_PostController extends Controller {
         );
     }
 
-
+    /**
+     * Deletes a blog post and redirects back to the manage posts page.
+     *
+     * Route: admin/blog/posts/delete/:id
+     *
+     * @param int $id The id of the blog post to delete.
+     */
     public static function delete($id)
     {
         Db::table('categories_posts')->where('post_id', '=', $id)->delete();
@@ -205,7 +203,6 @@ class Blog_Admin_PostController extends Controller {
 
         Url::redirect('admin/blog/posts');
     }
-
 
     /**
      * Gets categories formatted for html select box.
@@ -220,6 +217,5 @@ class Blog_Admin_PostController extends Controller {
 
         return $sortedCategories;
     }
-
 
 }
